LayoutTests/http/tests/inspector/network/network-disabling-check-no-memory-leak-expected.txt

CONSOLE MESSAGE: line 13: Done.
Tests that after disabling network domain, content saved on backend is removed. Bug 67995

resource.content after disabling network domain: null
